+++ src/Entity/OwnedCrypto.php
@@ -29,7 +29,7 @@ class OwnedCrypto
     private $plateforme;
 
     /**
-     * @ORM\Column(type="integer")
+     * @ORM\Column(type="float")
      */
     private $amount;
 
@@ -67,12 +67,12 @@ class OwnedCrypto
         return $this;
     }
 
-    public function getAmount(): ?int
+    public function getAmount(): ?float
     {
         return $this->amount;
     }
 
-    public function setAmount(int $amount): self
+    public function setAmount(float $amount): self
     {
         $this->amount = $amount;
